dom/mobileid/test/test_mobileid_no_permission.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for navigator.getMobileIdAssertion - No permission</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

ok(!("getMobileIdAssertion" in navigator),
   "navigator.getMobileIdAssertion should NOT exist");

SimpleTest.finish();
</script>
</pre>
</body>
</html>
